[Refactor] refact dsv4
