+++ vllm_metax/customized/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-License-Identifier: Apache-2.0
 # 2026 - Modified by MetaX Integrated Circuits (Shanghai) Co., Ltd. All Rights Reserved.
 
-from . import layers, ops, pluggable_layer, kernels  # noqa: F401
+from . import layers, ops, kernels  # noqa: F401
