--- vllm_metax/customized/pluggable_layer/__init__.py
@@ -1,5 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# 2026 - Modified by MetaX Integrated Circuits (Shanghai) Co., Ltd. All Rights Reserved.
-# from . import deepseek_v4_attention  # gdn_linear_attn  # noqa: F401
-# from . import gdn_linear_attn  # noqa: F401
-from . import gate_linear  # noqa: F401
