--- vllm_metax/model_executor/layers/mhc.py
@@ -1,526 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# 2026 - Modified by MetaX Integrated Circuits (Shanghai) Co., Ltd. All Rights Reserved.
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-import math
-from functools import cache
-from typing import TYPE_CHECKING
-
-import torch
-
-from vllm.platforms import current_platform
-from vllm.utils.import_utils import has_tilelang
-from vllm.utils.torch_utils import direct_register_custom_op
-
-# tilelang is only available on CUDA platforms
-if TYPE_CHECKING or current_platform.is_cuda_alike():
-    if not has_tilelang():
-        raise ImportError(
-            "tilelang is required for mhc but is not installed. Install it with "
-            "`pip install tilelang`."
-        )
-    import tilelang
-    import tilelang.language as T
-else:
-    tilelang = None  # type: ignore[assignment]
-    T = None  # type: ignore[assignment]
-
-
-@cache
-def compute_num_split(num_tokrns: int) -> int:
-    return 16 if num_tokrns >= 512 else 64
-
-
-def round_to_tf32(x: torch.Tensor) -> torch.Tensor:
-    return (x.view(torch.int32) + 0x1000).view(torch.float32)
-
-
-_PASS_CONFIGS = {
-    tilelang.PassConfigKey.TL_DISABLE_WGMMA: True,
-}
-
-
-@tilelang.jit(pass_configs=_PASS_CONFIGS)
-def _mhc_pre_norm_fn_fwd_mul(
-    mhc_mult3: int,
-    n_rms_group: int,
-    rms_group_size: int,
-    token_block: int = 32,
-    hidden_block: int = 256,
-) -> tilelang.JITKernel:
-    assert mhc_mult3 <= 32
-    num_tokens = T.dynamic("num_tokens")
-    assert rms_group_size % hidden_block == 0
-
-    @T.prim_func
-    def _mhc_pre_norm_fn_fwd_mul_kernel(
-        x: T.Tensor[(num_tokens, n_rms_group * rms_group_size), T.bfloat16],
-        fn: T.Tensor[(mhc_mult3, n_rms_group * rms_group_size), T.float32],
-        out: T.Tensor[(num_tokens, n_rms_group, mhc_mult3), T.float32],
-        sqrsum: T.Tensor[(num_tokens, n_rms_group), T.float32],
-    ) -> None:
-        _ = mhc_mult3
-        with T.Kernel(T.ceildiv(num_tokens, token_block), n_rms_group) as (
-            pid_x,
-            pid_y,
-        ):
-            out_frag = T.alloc_fragment((token_block, 32), T.float32)
-            sqrsum_part = T.alloc_fragment((token_block, 4), T.float32)
-            T.clear(out_frag)
-            T.clear(sqrsum_part)
-            for pz in T.Pipelined(rms_group_size // hidden_block, num_stages=1):
-                x_smem_16 = T.alloc_shared((token_block, hidden_block), T.bfloat16)
-                fn_smem = T.alloc_shared((32, hidden_block), T.float32)
-
-                T.annotate_layout(
-                    {x_smem_16: tilelang.layout.make_swizzled_layout(x_smem_16)}
-                )
-
-                T.copy(
-                    x[pid_x * token_block, pid_y * rms_group_size + pz * hidden_block],
-                    x_smem_16,
-                )
-                T.copy(fn[0, pid_y * rms_group_size + pz * hidden_block], fn_smem)
-
-                x_frag_16 = T.alloc_fragment((token_block, hidden_block), T.bfloat16)
-                T.copy(x_smem_16, x_frag_16)
-                x_frag = T.alloc_fragment((token_block, hidden_block), T.float32)
-                T.copy(x_frag_16, x_frag)
-
-                for jj in T.serial(hidden_block // 4):
-                    for i, j in T.Parallel(token_block, 4):
-                        sqrsum_part[i, j] += (
-                            x_frag[i, jj * 4 + j] * x_frag[i, jj * 4 + j]
-                        )
-
-                T.gemm(
-                    x_frag,
-                    fn_smem,
-                    out_frag,
-                    transpose_A=False,
-                    transpose_B=True,
-                    clear_accum=False,
-                )
-            sqrsum_l = T.alloc_fragment(token_block, T.float32)
-            T.reduce_sum(sqrsum_part, sqrsum_l)
-            for i in T.Parallel(token_block):
-                sqrsum[pid_x * token_block + i, pid_y] = sqrsum_l[i]
-            for i, j in T.Parallel(token_block, 32):
-                if j < 24:
-                    out[pid_x * token_block + i, pid_y, j] = out_frag[i, j]
-
-    return _mhc_pre_norm_fn_fwd_mul_kernel
-
-
-@tilelang.jit(
-    pass_configs={
-        tilelang.PassConfigKey.TL_DISABLE_WARP_SPECIALIZED: True,
-        tilelang.PassConfigKey.TL_PTXAS_REGISTER_USAGE_LEVEL: 10,
-        tilelang.PassConfigKey.TL_DISABLE_VECTORIZE_256: True,
-    },
-)
-def _mhc_pre_big_fuse(
-    hidden_size: int,
-    rms_eps: float,
-    mhc_pre_eps: float,
-    mhc_sinkhorn_eps: float,
-    mhc_post_mult_value: float,
-    sinkhorn_repeat: int,
-    n_splits: int = 16,
-    mhc_mult: int = 4,
-):
-    num_tokens = T.dynamic("num_tokens")
-    mhc_mult3 = mhc_mult * (2 + mhc_mult)
-    hidden_block = math.gcd(512, hidden_size)
-
-    @T.prim_func
-    def mhc_pre_big_fuse(
-        gemm_out_mul: T.Tensor[(n_splits, num_tokens, mhc_mult3), T.float32],
-        gemm_out_sqrsum: T.Tensor[(n_splits, num_tokens), T.float32],
-        mhc_scale: T.Tensor[(3,), T.float32],
-        mhc_base: T.Tensor[(mhc_mult3,), T.float32],
-        residual: T.Tensor[(num_tokens, mhc_mult, hidden_size), T.bfloat16],
-        # outputs
-        post_mix: T.Tensor[(num_tokens, mhc_mult), T.float32],
-        comb_mix: T.Tensor[(num_tokens, mhc_mult * mhc_mult), T.float32],
-        layer_input: T.Tensor[(num_tokens, hidden_size), T.bfloat16],
-    ) -> None:
-        with T.Kernel(num_tokens, threads=128) as pid:
-            ##################################################################
-            # _mhc_pre_norm_fn_fwd_norm
-            mixes_shared = T.alloc_shared(mhc_mult3, T.float32)
-            if T.get_thread_binding() < 64:
-                rms = T.alloc_fragment(1, T.float32)
-                mixes = T.alloc_fragment(mhc_mult3, T.float32)
-                T.clear(mixes)
-                rms[0] = 0
-                for i_split in T.serial(n_splits):
-                    rms[0] += gemm_out_sqrsum[i_split, pid]
-                rms[0] = T.rsqrt(rms[0] / (mhc_mult * hidden_size) + rms_eps)
-                for j in T.Parallel(mhc_mult3):
-                    mixes[j] = 0
-                    for i_split in T.serial(n_splits):
-                        mixes[j] += gemm_out_mul[i_split, pid, j]
-                    mixes[j] *= rms[0]
-                T.copy(mixes, mixes_shared, disable_tma=True)
-
-            if T.get_thread_binding() < 64:
-                ##################################################################
-                # _mhc_pre_split_mixes_fwd (post & comb)
-                cm = T.alloc_fragment((mhc_mult, mhc_mult), T.float32)
-                for j in T.Parallel(mhc_mult):
-                    post_mix[pid, j] = (
-                        T.sigmoid(
-                            mixes_shared[j + mhc_mult] * mhc_scale[1]
-                            + mhc_base[j + mhc_mult]
-                        )
-                        * mhc_post_mult_value
-                    )
-                for j, k in T.Parallel(mhc_mult, mhc_mult):
-                    cm[j, k] = (
-                        mixes_shared[j * mhc_mult + k + mhc_mult * 2] * mhc_scale[2]
-                        + mhc_base[j * mhc_mult + k + mhc_mult * 2]
-                    )
-
-                ##################################################################
-                # _mhc_sinkhorn_fwd
-                row_sum = T.alloc_fragment(mhc_mult, T.float32)
-                col_sum = T.alloc_fragment(mhc_mult, T.float32)
-
-                # comb = comb.softmax(-1) + eps
-                row_max = T.alloc_fragment(mhc_mult, T.float32)
-                T.reduce_max(cm, row_max, dim=1)
-                for j, k in T.Parallel(mhc_mult, mhc_mult):
-                    cm[j, k] = T.exp(cm[j, k] - row_max[j])
-                T.reduce_sum(cm, row_sum, dim=1)
-                for j, k in T.Parallel(mhc_mult, mhc_mult):
-                    cm[j, k] = cm[j, k] / row_sum[j] + mhc_sinkhorn_eps
-
-                # comb = comb / (comb.sum(-2) + eps)
-                T.reduce_sum(cm, col_sum, dim=0)
-                for j, k in T.Parallel(mhc_mult, mhc_mult):
-                    cm[j, k] = cm[j, k] / (col_sum[k] + mhc_sinkhorn_eps)
-
-                for _ in T.serial(sinkhorn_repeat - 1):
-                    # comb = comb / (comb.sum(-1) + eps)
-                    T.reduce_sum(cm, row_sum, dim=1)
-                    for j, k in T.Parallel(mhc_mult, mhc_mult):
-                        cm[j, k] = cm[j, k] / (row_sum[j] + mhc_sinkhorn_eps)
-
-                    # comb = comb / (comb.sum(-2) + eps)
-                    T.reduce_sum(cm, col_sum, dim=0)
-                    for j, k in T.Parallel(mhc_mult, mhc_mult):
-                        cm[j, k] = cm[j, k] / (col_sum[k] + mhc_sinkhorn_eps)
-
-                # save comb_mix to global memory
-                for j, k in T.Parallel(mhc_mult, mhc_mult):
-                    comb_mix[pid, j * mhc_mult + k] = cm[j, k]
-            else:
-                ##################################################################
-                # _mhc_pre_split_mixes_fwd (pre)
-                pre_mix_shared = T.alloc_shared(mhc_mult, T.float32)
-                for j in T.Parallel(mhc_mult):
-                    pre_mix_shared[j] = (
-                        T.sigmoid(
-                            mixes_shared[j] * mhc_scale[0] + mhc_base[j],
-                        )
-                        + mhc_pre_eps
-                    )
-                ###################################################################
-                # _mhc_pre_apply_mix_fwd
-                for i0_h in T.Pipelined(hidden_size // hidden_block, num_stages=2):
-                    xs = T.alloc_shared((mhc_mult, hidden_block), T.bfloat16)
-                    xl = T.alloc_fragment((mhc_mult, hidden_block), T.float32)
-                    T.copy(residual[pid, 0, i0_h * hidden_block], xs, disable_tma=True)
-                    T.copy(xs, xl, disable_tma=True)
-
-                    ol = T.alloc_fragment(hidden_block, T.float32)
-                    T.clear(ol)
-
-                    for i_mhc in T.serial(mhc_mult):
-                        pre = pre_mix_shared[i_mhc]
-                        for i1_h in T.Parallel(hidden_block):
-                            ol[i1_h] += pre * xl[i_mhc, i1_h]
-
-                    T.copy(ol, layer_input[pid, i0_h * hidden_block], disable_tma=True)
-
-    return mhc_pre_big_fuse
-
-
-def mhc_pre(
-    residual: torch.Tensor,
-    fn: torch.Tensor,
-    hc_scale: torch.Tensor,
-    hc_base: torch.Tensor,
-    rms_eps: float,
-    hc_pre_eps: float,
-    hc_sinkhorn_eps: float,
-    hc_post_mult_value: float,
-    sinkhorn_repeat: int,
-    n_splits: int = 1,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    """
-    Forward pass for mHC pre block.
-
-    Args:
-        residual: shape (..., hc_mult, hidden_size), dtype torch.bfloat16
-        fn: shape (hc_mult3, hc_mult * hidden_size), dtype torch.float32
-        hc_scale: shape (3,), dtype torch.float32
-        hc_base: shape (hc_mult3,), dtype torch.float32
-        rms_eps: RMS normalization epsilon
-        hc_pre_eps: pre-mix epsilon
-        hc_sinkhorn_eps: sinkhorn epsilon
-        hc_post_mult_value: post-mix multiplier value
-        sinkhorn_repeat: number of sinkhorn iterations
-        n_splits: split-k factor;
-
-    Returns:
-        post_mix: shape (..., hc_mult), dtype torch.float32
-        comb_mix: shape (..., hc_mult, hc_mult), dtype torch.float32
-        layer_input: shape (..., hidden_size), dtype torch.bfloat16
-    """
-
-    # Validate shapes
-    assert residual.dtype == torch.bfloat16
-    assert fn.dtype == torch.float32
-    assert hc_scale.dtype == torch.float32
-    assert hc_base.dtype == torch.float32
-
-    mhc_mult = residual.shape[-2]
-    hidden_size = residual.shape[-1]
-    mhc_mult2 = mhc_mult * mhc_mult
-    mhc_mult3 = mhc_mult * 2 + mhc_mult2
-
-    mhc_hidden_size = mhc_mult * hidden_size
-    assert fn.shape[0] == mhc_mult3
-    assert fn.shape[1] == mhc_hidden_size
-    assert hc_scale.shape == (3,)
-    assert hc_base.shape == (mhc_mult3,)
-
-    outer_shape = residual.shape[:-2]
-
-    residual_flat = residual.view(-1, mhc_mult, hidden_size)
-    num_tokens = residual_flat.shape[0]
-    fn_flat = fn  # noqa: F841
-
-    n_splits = compute_num_split(num_tokens)
-
-    post_mix = torch.empty(
-        num_tokens, mhc_mult, dtype=torch.float32, device=residual.device
-    )
-    comb_mix = torch.empty(
-        num_tokens, mhc_mult2, dtype=torch.float32, device=residual.device
-    )
-    layer_input = torch.empty(
-        num_tokens, hidden_size, dtype=torch.bfloat16, device=residual.device
-    )
-
-    gemm_out_mul = torch.empty(
-        n_splits, num_tokens, mhc_mult3, dtype=torch.float32, device=residual.device
-    )
-    gemm_out_sqrsum = torch.empty(
-        n_splits, num_tokens, dtype=torch.float32, device=residual.device
-    )
-
-    from vllm_metax.utils.deep_gemm import tf32_hc_prenorm_gemm
-
-    tf32_hc_prenorm_gemm(
-        residual_flat.view(num_tokens, mhc_mult * hidden_size),
-        fn_flat,
-        gemm_out_mul,
-        gemm_out_sqrsum,
-        n_splits,
-    )
-    # END of TileLang implementation of pre-norm-fn forward matmul
-
-    _mhc_pre_big_fuse(
-        hidden_size,
-        rms_eps,
-        hc_pre_eps,
-        hc_sinkhorn_eps,
-        hc_post_mult_value,
-        sinkhorn_repeat,
-        n_splits=n_splits,
-        mhc_mult=mhc_mult,
-    )(
-        gemm_out_mul,
-        gemm_out_sqrsum,
-        hc_scale,
-        hc_base,
-        residual_flat,
-        post_mix,
-        comb_mix,
-        layer_input,
-    )
-
-    post_mix = post_mix.view(*outer_shape, mhc_mult, 1)
-    comb_mix = comb_mix.view(*outer_shape, mhc_mult, mhc_mult)
-    layer_input = layer_input.view(*outer_shape, hidden_size)
-
-    return post_mix, comb_mix, layer_input
-
-
-def _mhc_pre_fake(
-    residual: torch.Tensor,
-    fn: torch.Tensor,
-    hc_scale: torch.Tensor,
-    hc_base: torch.Tensor,
-    rms_eps: float,
-    hc_pre_eps: float,
-    hc_sinkhorn_eps: float,
-    hc_post_mult_value: float,
-    sinkhorn_repeat: int,
-    n_splits: int = 1,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    hc_mult = residual.shape[-2]
-    hidden_size = residual.shape[-1]
-    outer_shape = residual.shape[:-2]
-
-    # Create empty tensors with correct shapes for meta device / shape inference
-    post_mix = torch.empty(
-        *outer_shape,
-        hc_mult,
-        1,
-        dtype=torch.float32,
-        device=residual.device,
-    )
-    comb_mix = torch.empty(
-        *outer_shape,
-        hc_mult,
-        hc_mult,
-        dtype=torch.float32,
-        device=residual.device,
-    )
-    layer_input = torch.empty(
-        *outer_shape,
-        hidden_size,
-        dtype=torch.bfloat16,
-        device=residual.device,
-    )
-
-    return post_mix, comb_mix, layer_input
-
-
-@tilelang.jit(
-    pass_configs={
-        tilelang.PassConfigKey.TL_DISABLE_WARP_SPECIALIZED: True,
-        tilelang.PassConfigKey.TL_PTXAS_REGISTER_USAGE_LEVEL: 10,
-        tilelang.PassConfigKey.TL_DISABLE_VECTORIZE_256: True,
-    },
-)
-def _mhc_post_fwd(
-    mhc: int, hidden: int, n_thr: int = 128, h_blk: int = 1024
-) -> tilelang.JITKernel:
-    n = T.dynamic("num_tokens")
-    h = hidden
-
-    h_blk = math.gcd(hidden, h_blk)
-
-    @T.prim_func
-    def _mhc_post_fwd_kernel(
-        a: T.Tensor[(n, mhc, mhc), T.float32],
-        b: T.Tensor[(n, mhc, h), T.bfloat16],
-        c: T.Tensor[(n, mhc), T.float32],
-        d: T.Tensor[(n, h), T.bfloat16],
-        x: T.Tensor[(n, mhc, h), T.bfloat16],
-    ) -> None:
-        with T.Kernel(n, threads=n_thr) as pid_n:
-            x_shared = T.alloc_shared((mhc, h_blk), T.bfloat16)
-            b_shared = T.alloc_shared((mhc, h_blk), T.bfloat16)
-            d_shared = T.alloc_shared(h_blk, T.bfloat16)
-
-            x_local = T.alloc_fragment((mhc, h_blk), T.float32)
-            b_local = T.alloc_fragment((mhc, h_blk), T.float32)
-            d_local = T.alloc_fragment(h_blk, T.float32)
-
-            a_local = T.alloc_fragment((mhc, mhc), T.float32)
-            c_local = T.alloc_fragment(mhc, T.float32)
-            T.copy(a[pid_n, 0, 0], a_local)
-            T.copy(c[pid_n, 0], c_local)
-
-            for i0_h in T.Pipelined(T.ceildiv(h, h_blk), num_stages=2):
-                T.copy(b[pid_n, 0, i0_h * h_blk], b_shared, disable_tma=True)
-                T.copy(d[pid_n, i0_h * h_blk], d_shared, disable_tma=True)
-
-                T.copy(b_shared, b_local)
-                T.copy(d_shared, d_local)
-                for i_mhco, i1_h in T.Parallel(mhc, h_blk):
-                    x_local[i_mhco, i1_h] = c_local[i_mhco] * d_local[i1_h]
-                    for i_mhci in T.serial(mhc):
-                        x_local[i_mhco, i1_h] += (
-                            a_local[i_mhci, i_mhco] * b_local[i_mhci, i1_h]
-                        )
-                T.copy(x_local, x_shared)
-
-                T.copy(x_shared, x[pid_n, 0, i0_h * h_blk], disable_tma=True)
-
-    return _mhc_post_fwd_kernel
-
-
-def mhc_post_fwd(
-    x: torch.Tensor,
-    residual: torch.Tensor,
-    post_layer_mix: torch.Tensor,
-    comb_res_mix: torch.Tensor,
-    out: torch.Tensor | None = None,
-) -> torch.Tensor:
-    num_tokens, mhc, hidden = residual.shape
-
-    assert x.dtype == torch.bfloat16, f"{x.dtype=}"
-    assert residual.dtype == torch.bfloat16, f"{residual.dtype=}"
-    assert post_layer_mix.dtype == torch.float32, f"{post_layer_mix.dtype=}"
-    assert comb_res_mix.dtype == torch.float32, f"{comb_res_mix.dtype=}"
-    assert x.shape == (num_tokens, hidden), f"{x.shape=}"
-    assert post_layer_mix.shape == (num_tokens, mhc, 1), f"{post_layer_mix.shape=}"
-    assert comb_res_mix.shape == (num_tokens, mhc, mhc), f"{comb_res_mix.shape=}"
-
-    residual = residual.contiguous()
-    assert x.is_contiguous()
-    assert post_layer_mix.is_contiguous()
-    assert comb_res_mix.is_contiguous()
-
-    if out is None:
-        out = torch.empty_like(residual)
-
-    kernel = _mhc_post_fwd(mhc, hidden)
-    kernel(
-        comb_res_mix,
-        residual,
-        post_layer_mix.squeeze(-1),
-        x,
-        out,
-    )
-    return out
-
-
-def mhc_post(
-    x: torch.Tensor,
-    residual: torch.Tensor,
-    post_layer_mix: torch.Tensor,
-    comb_res_mix: torch.Tensor,
-) -> torch.Tensor:
-    out = torch.empty_like(residual)
-
-    return mhc_post_fwd(x, residual, post_layer_mix, comb_res_mix, out)
-
-
-def _mhc_post_fake(
-    x: torch.Tensor,
-    residual: torch.Tensor,
-    post_layer_mix: torch.Tensor,
-    comb_res_mix: torch.Tensor,
-) -> torch.Tensor:
-    return torch.empty_like(residual)
-
-
-direct_register_custom_op(
-    op_name="mx_mhc_pre",
-    op_func=mhc_pre,
-    mutates_args=[],
-    fake_impl=_mhc_pre_fake,
-)
-direct_register_custom_op(
-    op_name="mx_mhc_post",
-    op_func=mhc_post,
-    mutates_args=[],
-    fake_impl=_mhc_post_fake,
-)
