+++ vllm_metax/models/deepseek_v4/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: Apache-2.0
+# 2026 - Modified by MetaX Integrated Circuits (Shanghai) Co., Ltd. All Rights Reserved.
+
+from .model import DeepseekV4ForCausalLM  # type: ignore[assignment]
+from .mtp import DeepSeekV4MTP  # type: ignore[assignment]
+
+__all__ = [
+    "DeepSeekV4MTP",
+    "DeepseekV4ForCausalLM",
+]
