+++ vllm_metax/models/deepseek_v4/ops/__init__.py
@@ -1,0 +1,12 @@
+# SPDX-License-Identifier: Apache-2.0
+# 2026 - Modified by MetaX Integrated Circuits (Shanghai) Co., Ltd. All Rights Reserved.
+
+from .fused_indexer_q import fused_indexer_q_rope_int8_quant
+from .cache_utils import gather_k_cache
+from .fused_inv_rope_quant import inv_rope
+
+__all__ = [
+   "fused_indexer_q_rope_int8_quant",
+   "gather_k_cache",
+   "inv_rope",
+]
