+++ vllm_metax/patch/bugfix/deepseek_v4/__init__.py
@@ -1,3 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 # 2026 - Modified by MetaX Integrated Circuits (Shanghai) Co., Ltd. All Rights Reserved.
 from . import parallel_state  # noqa: F401
+# ---------------------------
+# Note: enable for dsv4 branch with torch2.8
+# from . import torch_fix
