--- vllm_metax/v1/attention/ops/deepseek_v4_ops/__init__.py
@@ -1,9 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# 2026 - Modified by MetaX Integrated Circuits (Shanghai) Co., Ltd. All Rights Reserved.
-from .cache_utils import (
-    inv_rope,
-    fused_deepseek_v4_qnorm_rope_kv_rope_insert,
-    fused_indexer_q_rope_int8_quant,
-    gather_k_cache,
-    gather_k_cache_bf16,
-)
